+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - Added support to Logical switch group actions
   - Added support to Logical switch actions
   - Added support to Firmware bundle actions
+  - Added diff output for updating resources
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0